Separate out USB panel into a page.
