+++ InfoPanel/App.xaml.cs
@@ -102,6 +102,8 @@ namespace InfoPanel
            services.AddScoped<SettingsViewModel>();
            services.AddScoped<Views.Pages.UpdatesPage>();
            services.AddScoped<UpdatesViewModel>();
+           services.AddScoped<Views.Pages.UsbPanelsPage>();
+           services.AddScoped<UsbPanelsViewModel>();
 
            // Configuration
            //services.Configure<AppConfig>(context.Configuration.GetSection(nameof(AppConfig)));
