+++ InfoPanel/ViewModels/SettingsViewModel.cs
@@ -14,18 +14,6 @@ namespace InfoPanel.ViewModels
         public float Value { get; set; }
     }
 
-    public enum LCD_ROTATION
-    {
-        [Description("No rotation")]
-        RotateNone = 0,
-        [Description("Rotate 90°")]
-        Rotate90FlipNone = 1,
-        [Description("Rotate 180°")]
-        Rotate180FlipNone = 2,
-        [Description("Rotate 270°")]
-        Rotate270FlipNone = 3,
-    }
-
     public partial class SettingsViewModel : ObservableObject, INavigationAware
     {
         private ObservableCollection<string> _comPorts = [];
@@ -39,26 +27,13 @@ namespace InfoPanel.ViewModels
                 new UiScaleOption { Display = "120%", Value = 1.2f }
             ];
 
-        public ObservableCollection<LCD_ROTATION> RotationValues { get; set; }
-
         public SettingsViewModel()
         {
-            RotationValues = new ObservableCollection<LCD_ROTATION>(Enum.GetValues(typeof(LCD_ROTATION)).Cast<LCD_ROTATION>());
         }
 
         public ObservableCollection<string> ComPorts
         {
             get { return _comPorts; }
-        }
-
-        public ObservableCollection<BeadaPanelDevice> RuntimeBeadaPanelDevices
-        {
-            get { return ConfigModel.Instance.Settings.BeadaPanelDevices; }
-        }
-
-        public ObservableCollection<TuringPanelDevice> RuntimeTuringPanelDevices
-        {
-            get { return ConfigModel.Instance.Settings.TuringPanelDevices; }
         }
 
         public void OnNavigatedFrom()
